+++ R/preparation.R
@@ -18,7 +18,7 @@
 }
 
 .onAttach <- function(libname, pkgname) {
-  packageStartupMessage("Good morning star-shine, cNORM says 'Hello!'")
+  packageStartupMessage("Good morning star-shine!\ncNORM is free software. Please report bugs: https://github.com/WLenhard/cNORM/issues")
 }
 
 
@@ -40,9 +40,7 @@
 #' observed raw scores, otherwise, ranking is done by group (default)
 #' @param weights Vector or variable name in the dataset with weights for each individual case. It can be used
 #' to compensate for moderate imbalances due to insufficient norm data stratification. Weights should be numerical
-#' and positive.
-#' Please note, that this feature is currently EXPERIMENTAL and subject to ongoing work! Precision of weighting increases
-#' with sample size. On the other hand, in large samples, it is easy to stratificate and then weighting is not needed anymore.
+#' and positive. Please use the 'computeWeights' function for this purpose.
 #' @param scale type of norm scale, either T (default), IQ, z or percentile (= no
 #' transformation); a double vector with the mean and standard deviation can as well,
 #' be provided f. e. c(10, 3) for Wechsler scale index point
@@ -176,14 +174,12 @@ prepareData <- function(data = NULL, group = "group", raw = "raw", age = "group"
 #'
 #' @param data data.frame with norm sample data. If no data.frame is provided, the raw score
 #' and group vectors are directly used
-#' @param group name of the grouping variable (default 'group')  or numeric vector, e. g. grade, setting
+#' @param group name of the grouping variable (default 'group') or numeric vector, e. g. grade, setting
 #' group to FALSE cancels grouping (data is treated as one group)
 #' @param raw name of the raw value variable (default 'raw') or numeric vector
 #' @param weights Vector or variable name in the dataset with weights for each individual case. It can be used
 #' to compensate for moderate imbalances due to insufficient norm data stratification. Weights should be numerical
-#' and positive.
-#' Please note, that this feature is currently EXPERIMENTAL and subject to ongoing work! Precision of weighting increases
-#' with sample size. On the other hand, in large samples, it is easy to stratificate and then weighting is not needed anymore.
+#' and positive.  Please use the 'computeWeights' function for this purpose.
 #' @param method Ranking method in case of bindings, please provide an index,
 #' choosing from the following methods: 1 = Blom (1958), 2 = Tukey (1949),
 #' 3 = Van der Warden (1952), 4 = Rankit (default), 5 = Levenbach (1953),
@@ -217,7 +213,7 @@ prepareData <- function(data = NULL, group = "group", raw = "raw", age = "group"
 #' d <- computePowers(d)
 #' m <- bestModel(d)
 #' rawTable(0, m) # please use an arbitrary value for age when generating the tables
-#' @seealso rankBySlidingWindow, computePowers
+#' @seealso rankBySlidingWindow, computePowers, computeWeights, weighted.rank
 #' @export
 #' @family prepare
 rankByGroup <-
@@ -270,9 +266,6 @@ rankByGroup <-
 
     weighting <- NULL
     if(!is.null(weights)){
-        # message("Weighting is currently not working in rankByGroup. Proceeding without weighting.")
-        # weights <- NULL
-
       if(is.character(weights)){
 
         if(!(weights %in% colnames(d))){
@@ -351,6 +344,7 @@ rankByGroup <-
 
     if (is.null(covariate)) {
       if (typeof(group) == "logical" && !group) {
+        cat("No grouping variable specified. Ranking without grouping ...")
         d$percentile <- (weighted.rank(sign * (d[, raw]), weights = weighting) + numerator[method]) / (length(d[, raw]) + denominator[method])
         if (descriptives) {
           d$n <- length(d[, raw])
@@ -388,7 +382,7 @@ rankByGroup <-
       warning("Using covariates is an EXPERIMENTAL feature in this package currently.")
 
       if (typeof(group) == "logical" && !group) {
-
+            cat("No grouping variable specified. Ranking without grouping ...")
             d$percentile <- ave(d[, raw], d[, covariate], FUN = function(x) {
               (weighted.rank(sign * x, weights = weighting) + numerator[method]) / (length(x) + denominator[method])
             })
@@ -525,9 +519,7 @@ rankByGroup <-
 #' @param width the width of the sliding window
 #' @param weights Vector or variable name in the dataset with weights for each individual case. It can be used
 #' to compensate for moderate imbalances due to insufficient norm data stratification. Weights should be numerical
-#' and positive.
-#' Please note, that this feature is currently EXPERIMENTAL and subject to ongoing work! Precision of weighting increases
-#' with sample size. On the other hand, in large samples, it is easy to stratificate and then weighting is not needed anymore.
+#' and positive. It can be resource intense when applied to the sliding window. Please use the 'computeWeights' function for this purpose.
 #' @param method Ranking method in case of bindings, please provide an index,
 #' choosing from the following methods: 1 = Blom (1958), 2 = Tukey (1949),
 #' 3 = Van der Warden (1952), 4 = Rankit (default), 5 = Levenbach (1953),
@@ -564,7 +556,7 @@ rankByGroup <-
 #' data.elfe <- rankByGroup(elfe, group = "group")
 #' mean(data.elfe$normValue - data.elfe2$normValue)
 #' }
-#' @seealso rankByGroup, computePowers
+#' @seealso rankByGroup, computePowers, computeWeights, weighted.rank, weighted.quantile
 #' @export
 #' @family prepare
 rankBySlidingWindow <- function(data = NULL,